do not import vscode in webview
